+++ src/main/java/com/balugaq/netex/utils/ImageUtil.java
@@ -1,4 +1,4 @@
-package com.balugaq.netex.api.texture;
+package com.balugaq.netex.utils;
 
 import java.awt.image.BufferedImage;
 import java.io.IOException;
